+++ internal/domain/model/product_boms.go
@@ -2,15 +2,15 @@ package model
 
 // Component for BOM - represents one component needed
 type BomComponent struct {
-	ComponentProductID int     `json:"component_product_id" binding:"required"` // ID sản phẩm nguyên liệu
-	Quantity           float64 `json:"quantity" binding:"required"`             // Số lượng nguyên liệu cần thiết
+	ComponentProductID int `json:"component_product_id" binding:"required"` // ID sản phẩm nguyên liệu
+	Quantity           int `json:"quantity" binding:"required"`             // Số lượng nguyên liệu cần thiết
 }
 
 // Component with full product info for response
 type BomComponentResponse struct {
 	ID                 int             `json:"id"`                          // ID của BOM entry
 	ComponentProductID int             `json:"component_product_id"`        // ID sản phẩm nguyên liệu
-	Quantity           float64         `json:"quantity"`                    // Số lượng nguyên liệu cần thiết
+	Quantity           int             `json:"quantity"`                    // Số lượng nguyên liệu cần thiết
 	ComponentProduct   *ProductBomInfo `json:"component_product,omitempty"` // Thông tin sản phẩm nguyên liệu
 }
 
@@ -52,22 +52,22 @@ type ProductBomInfo struct {
 
 // Material requirement calculation request
 type CalculateMaterialRequirementsRequest struct {
-	ParentProductID int     `json:"parent_product_id" binding:"required"` // ID sản phẩm thành phẩm
-	Quantity        float64 `json:"quantity" binding:"required,gt=0"`     // Số lượng sản phẩm cần sản xuất
+	ParentProductID int `json:"parent_product_id" binding:"required"` // ID sản phẩm thành phẩm
+	Quantity        int `json:"quantity" binding:"required,gt=0"`     // Số lượng sản phẩm cần sản xuất
 }
 
 // Raw material requirement
 type MaterialRequirement struct {
 	ProductID        int             `json:"product_id"`        // ID nguyên liệu
 	Product          *ProductBomInfo `json:"product"`           // Thông tin nguyên liệu
-	RequiredQuantity float64         `json:"required_quantity"` // Tổng số lượng cần thiết
+	RequiredQuantity int             `json:"required_quantity"` // Tổng số lượng cần thiết
 }
 
 // Material requirements calculation response
 type MaterialRequirementsResponse struct {
 	ParentProductID      int                   `json:"parent_product_id"`     // ID sản phẩm thành phẩm
 	ParentProduct        *ProductBomInfo       `json:"parent_product"`        // Thông tin sản phẩm thành phẩm
-	RequestedQuantity    float64               `json:"requested_quantity"`    // Số lượng sản phẩm được yêu cầu
+	RequestedQuantity    int                   `json:"requested_quantity"`    // Số lượng sản phẩm được yêu cầu
 	MaterialRequirements []MaterialRequirement `json:"material_requirements"` // Danh sách nguyên liệu và số lượng cần thiết
 	TotalMaterials       int                   `json:"total_materials"`       // Tổng số loại nguyên liệu
 }
